// AcsStatsAngular/ClientApp/src/app/modules/batting-records/models/batting-overall.model.ts
export interface BattingCareerRecordDto {
  name: string;
  sortNamePart: string;
  team: string;
  opponents: string;
  year: string;
  matches: number;
  innings: number;
  ground: string;
  countryName: string;
  runs: number;
  notOuts: number;
  highestScore: number;
  notOut: boolean;
  avg: number;
  hundreds: number;
  fifties: number;
  ducks: number;
  fours: number;
  sixes: number;
  balls: number;
}