// AcsStatsAngular/ClientApp/src/app/models/sortorder.model.ts
export enum SortOrder {
  undefined = 0,
  name = 1,
  team = 2,
  opponents = 3,
  runs = 4,
  totals = 4,
  wickets = 5,
  balls = 6,
  season = 7,
  matches = 8,
  wins = 9,
  lost = 10,
  tied = 11,
  drawn = 12,
  byes = 13,
  legByes = 14,
  wides = 15,
  noBalls = 16,
  matchStartDate = 17,
  victoryMargin = 18,
  seriesDate = 19,
  ground = 20,
  hostCountry = 21,
  matchStartYear = 22,
  penalties = 23,
  extras = 24,
  percentage = 25,
  dismissals = 26,
  caught = 27,
  stumpings = 28,
  cwk = 29,
  cf = 30,
  bestDismissals = 31,
  innings = 32,
  fours = 33,
  sixes = 34,
  inningsNumber = 35,
  bat1 = 36,
  bat2 = 37,
  minutes = 38,
  notOuts = 39,
  hs = 40,
  avg = 41,
  fifties = 42,
  hundreds = 43,
  ducks = 44,
  bbi = 45,
  bbm = 46,
  maidens = 47,
  dots = 48,
  tenfor = 49,
  fivefor = 50,
  econ = 51,
  strikeRate = 52,
  ballsPerOver = 53,
}